chore(ds): add useGraphQLQuery hooks
